requirements.txt
aiogram
aiohttp
aiofiles
aiocryptopay
aiogram_dialog
APScheduler
asyncpg
environs
nats-py
ormsgpack
greenlet
pydantic
SQLAlchemy
pytz
openai
yookassa-async

